querystring, url: factor out common parts
