+++ lib/querystring.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const { Buffer } = require('buffer');
+const { StorageObject, hexTable } = require('internal/querystring');
 const QueryString = module.exports = {
   unescapeBuffer,
   // `unescape()` is a JS global, so we need to use a different local name
@@ -14,13 +16,6 @@ const QueryString = module.exports = {
   parse,
   decode: parse
 };
-const Buffer = require('buffer').Buffer;
-
-// This constructor is used to store parsed query string values. Instantiating
-// this is faster than explicitly calling `Object.create(null)` to get a
-// "clean" empty object (tested with v8 v4.9).
-function ParsedQueryString() {}
-ParsedQueryString.prototype = Object.create(null);
 
 const unhexTable = [
   -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, // 0 - 15
@@ -115,10 +110,6 @@ function qsUnescape(s, decodeSpaces) {
   }
 }
 
-
-const hexTable = [];
-for (var i = 0; i < 256; ++i)
-  hexTable[i] = '%' + ((i < 16 ? '0' : '') + i.toString(16)).toUpperCase();
 
 // These characters do not need escaping when generating query strings:
 // ! - . _ ~
@@ -263,7 +254,7 @@ const defEqCodes = [61]; // =
 
 // Parse a key/val string.
 function parse(qs, sep, eq, options) {
-  const obj = new ParsedQueryString();
+  const obj = new StorageObject();
 
   if (typeof qs !== 'string' || qs.length === 0) {
     return obj;
